use libvirt via libvirt-go
The libvirt-go module is not very easy to use in unit tests. There is a
"test:///default" connection URI which could be used for this purpose,
but it does not support certain functions that we need such as volume
upload. So we use mocking instead. However, mocking is messy because the
Connect struct has methods which return other structs such as
StoragePool which also need to be mocked. Hence libvirt.Connect cannot
directly fulfill the interface that we require and we have to wrap it.

As an additional complication, I wanted to keep the mocks in their own
directory. This puts them in their own package, which means they can't
refer back to the main package to use the other interfaces. I solved
this by moving the interfaces into a package of their own.