switch to fake libvirt instead of mockery mock
Previously the tests were hard to understand because they had to mock
complicated sequences of libvirt interactions. In addition, the mock
implementation generated by mockery was difficult to use when method
calls might return different values after other interactions.

Now, instead of mocking the interface with a generic generated mock, we
use a dummy libvirt implementation with some simplifications that are
appropriate for our requirements. This makes the tests easier to
understand.