vm: use VMConfig
Let's use VMConfig if we have it. Now we already got parameter lists
that look like this:

ip, err := v.createVM(sp, vmConfig.VMName, vmConfig.VMID,
  vmConfig.MemoryKiB)

While it is nice to see all the required paramters, I also think this is
more for historic reasons when there were less options passed around.
For documentation purposes we can assign the required parts of the
config to local variables at the beginning of the function as I did
here. Better fix that now than later.