libvirt: use host-passthrough CPU mode
This seems to be the most sensible option for our particular use case.
Pretty much the only downside of host-passthrough is that migrating VMs
becomes more dangerous (because the CPU needs to be *exactly* the same),
but we don't currently want to migrate VMs anyways, so that is a
non-issue for us.

The manual also states that:

    [...] the downside of this mode is that the guest environment cannot
    be reproduced on different hardware. Thus, if you hit any bugs,
    you are on your own.

So if we hit any problems with that we should probably think about
making this configurable or using host-model as a "safer" alternative.