provision: add docker and shell/ssh
This is a minimal version, but already shows to be pretty useful.

In the longer run we need so separate docker and ssh provisioning in
ImageBuild. There is no need to require docker if it is shell/ssh
provisioning only. But this minimal version keeps the interface.
We can fix that later.