provisioning: add rsync provisioning step
The rsync provisioning step allows users to copy files from the host
machine to the guest VM(s).

It does so by passing the respective paths to the `rsync` utility, which
is expected to be installed on the host as well as on the guest machines.

There is a new interface, NetworkCopier, which describes an object that
can copy files over the network. Also, an implementation of this
interface using rsync is provided.