cmd: if config file is not found, write defaults
If we are trying to load a config file that does not exist, write a
default config file to that location, with all keys set to their default
value. This offers the additional benefit of having inline documentation in
the default config file.

While at it, group all the config related functions into a common file.