Dockerfile: add file
This allows running virter as a docker container. Unfortunately virter
simply is not that trivial to run in a container as it will need at
least a connection to libvirt's socket. If one wants to use docker
provisioning, the docker socket also needs to be mapped to the
container. But having the mappings in place, this works pretty nicely,
and it can always be put behind an alias.

Example:
$ docker run -it --rm \
  -v ~/.config/virter:/root/.config/virter:ro
  -v /var/run/libvirt/libvirt-sock:/var/run/libvirt/libvirt-sock \
  virter vm ssh virter-centos-7