vm run: remove broken attempt to insert IP into hosts
This was broken because the IP address was never passed to the user
data. Remove it because it is not really compatible with the `image
build` workflow. The build would add an IP address to /etc/hosts which
is no longer valid after the build is complete. When using this image,
this could cause confusion.