iso/cloud-init: remove genisoimage dependency
There is a nice library that works for us that looks clean and does not
introduce any other external dependencies. It is a bit uncommon in the
sense that it requires the destination to fulfil the WriteAt()
interface, which standard buffers don't. There is already prior art
under a Apache2 license, which I just stripped down a bit.