cmd: implement image ls command (--available only)
When multiple image registry sources are being used, it can be tricky to
figure out what images are actually available to pull.
For this, implement `virter image ls --available`. This reads in the image
registry files and prints all image names along with their URLs.

The awkward part is that we already have another feature on our roadmap
which is supposed to list all the images available locally (i.e. all
images that can be used in `vm run` or similar). So we decided that this
should be accessible using `virter image ls --available`, and the other
one should just be `virter image ls`.
For now, just print an error message for the not-yet-implemented `image ls`.