cmd: source registry from default and user-defined file
Instead of just getting the registry data from the user-supplied
images.toml file, also read it from built-in registry.

If the built-in registry file does not exist, it is created first.

This also means that the image.registry configuration option as well as
the assets/images.toml file are now no longer needed and can be removed.