vm: store cloud-config templates in code instead of files
Instead of reading user-data and meta-data from the assets/ folder
(which we are trying to get rid of), just store them in code. They are
small enough to not be too annoying to have in the Go code.

Also take this opportunity to move some cloud-config related stuff out
of vm.go since that file is getting pretty huge.