rsync: be louder when source does not match anything
Originally I thought that this was more of a "nice to know" kind of
condition. However, now that I have used this provisioning step for a
few provisioners, I have noticed that it is pretty much always going to
be a user error (typo, some NFS share not mounted, whatever).

To make things worse, when an error like this occurs, absolutely
*nothing* is printed, which is really confusing. So, I would suggest to
bump this message to a warning, so that is nice and obvious from the
logs that something is probably wrong.