replace TCP ping with SSH ping
It turns out that some older distributions like CentOS 7 open the SSH
port before the SSH server is actually ready to start accepting
connections.  This causes problems when attempting to connect to a VM
immediately after it has booted, which is exactly what the image build
process does.

Try to open a real SSH connection instead of just a TCP connection to
handle this.

We introduce a ShellClient abstraction so that the SSH connection can be
replaced by a mock implementation for tests.