VMRun: check as many preconditions as possible
Rolling back a VM is surprisingly difficult if one wants to make sure to
not delete parts (e.g., IP, an image) of an existing/valid VM.

So the approach here is to check all the "objects" the new VM would
create don't exist yet. If one exists, we fail before we create the
others.

After these checks we could just VMRm the VM, but:
- as is it should cover like 99% of all use cases
- we are sure we don't delete anything valid because we don't delete...

There are obvious races, but this is unfortunately not the first one in
this path. We should fix that by also flock()-ing.