build: reset /etc/machine-id
Without that a machine-id is generated that is then not reset when we
"virter vm run" machines. Just inject a build step resetting it.

"truncate -c" should make sure that we don't create the file if it did
not exist, and it returns 0, so we don't need "|| true".

"-c", "-s" are even implemented in busybox truncate, so we can be sure
it works.