cmd: add --disk flag to vm run
Add a new flag, --disk (or -d for short), which can be used to attach
disks of arbitrary size to the VM.

The flag is a string in the format "key=value,key2=value2". Formally,
you could say these are "comma-separated key-value pairs". The --disk
parameter can be specified multiple times to attach multiple disks.

This change brings quite a bit of internal restructuring. Most notably,
previously we had a fixed set of disks for each VM: the boot disk, the
cidata CDROM, and the "scratch" disk (with a fixed size).
When deleting a VM, these could be deleted since they would always have
deterministic names.

This now changes, because the disk names are set by the user and are
therefore arbitrary; we have to read them back from the libvirt domain XML
definition if we want to delete them.