libvirtxml: fixup broken nil check
Fixup for recent commit 51993f87de57 ("cmd: add --disk flag to vm run").

Well, we did check for nil there, and we even logged that we were going
to skip this entry. But apparently I forgot to actually put the
`continue` there, which made it access the nil pointer anyways...