vm run: allow starting multiple VMs
It is really common to start multiple VMs at once (for example, for a
cluster). Implement this using a --count flag which just counts up from
the ID given with --id. The VM name is either derived from the image
name ("image"-"id"), from the name given with --name and the ID if count
is > 1 ("name"-"id"), or only from the name given with --name if
count == 1 ("name").