provisioning: display VM name in stdout/err stream
It would make sense to see which VM is producing the SSH output we are
seeing, so instead of "SSH stdout/stderr", prefix the line with the VM
name. This also has the added benefit of being more concise most of the
time, so the log lines aren't as long.

In order to still differentiate between stdout and stderr, we color
stderr red (which matches what other tools such as vagrant and Docker
do).

For Docker provisioning, we just print "Docker" instead of the VM name
(because that targets multiple VMs).