vm: let libvirt choose the MAC address for us
There appear to be some weird conflicts within libvirt when creating a
VM with a MAC address that already has a DHCP lease. To work around
this, make sure to always get a fresh MAC address by letting libvirt
choose one for us.

Also adapt the tests; the fake libvirt implementation should "allocate"
a MAC address for us on DomainDefineXML. We also don't want to assert
the correctness of the returned MAC address anymore, because we don't
deal with that value anymore.