vm rm: try releasing the VMs DHCP lease
We have seen some strange effects when IP/MAC addresses are reused. A
solution for this problem is to get rid of the DHCP lease when
destroying the VM.
In order to do that, we have to run dhcp_release, which can only run as
root. We try running it with sudo when removing the machine. If it
doesn't work, we don't complain either.

Note that this requires passwordless sudo for the executing user, at
least for the dhcp_release command. Something along the lines of:

    %libvirt ALL=(ALL) NOPASSWD: /usr/bin/dhcp_release

To allow all users in the libvirt group to execute it.