work around a deadlock when waiting for VM shutdown
When building images in parallel – using vmshed in this case – virter
would sometimes lock up when committing the image. It would always hang
when submitting the libvirt API request for the `StorageVolLookupByName`
call in `vmRmExceptBoot`.

I also noticed that the goroutine in the `LifecycleEvents` call never
seems to return; and in the cases where virter deadlocked, a lifecycle
event was received after we were no longer actively listening for them.
I don't know exactly why this trips up libvirt, but it does.

Thus, switch to a primitive polling loop for tracking the VM's shutdown
progress.

This also means that we can get rid of LifecycleEvents altogether, since
this was the only user.