add static DHCP option
This introduces a new mode that can be toggled via a config option. When
this mode is used, no DHCP host entries are changed when creating or
removing VMs. This is useful because network updates can disrupt the
network connectivity of other VMs. Using this mode, the problematic
network updates can be better controlled.