centralize handling of ssh keys and configuration
In preparation for the upcoming rework of ssh host keys and ssh authentication
in containers, this commit centralized ssh key handling to a single struct
attached to the virter instance.

This commit adds `pkg/sshkeys` and `KeyStore` interface, which:
* Loads private and public key from FS
* Generates a private/public pair if it does exist
* Provides methods to access the private key as:
  * Path, used by rsync
  * Bytes, used in go's ssh client and for containers
* Provides mthods to access the public key as:
  * Bytes, used to configure the VM's authorized_keys

The KeyStore is attached to the Virter struct instead of manually
adding the required form of private/public key to each method
that may of may not need it. As such, the `VirterConnect` function
was renamed `InitVirter` to reflect that setting up the libvirt
connection is not the only thing it does.

As a consequence, `VmConfig` now expects only `ExtraSSHPublicKeys`
instead of all public keys. The virter public key is automatically
added to the VM's authorized_keys.

Another change: the `KeyStore` is passed to the NetworkCopier.Copy interface.
This was done as because the KeyStore is only available to the methods
on the `Virter` struct, which just call an already constructed NetworkCopier.