ci: fetch target branch before lint run
gitlab only fetches relevant refs, i.e. the current branch. This means
that if a runner never runs a master branch job, it won't have the master
branch available to compare. It also means that the comparison is against
potentially old commits instead of the current branch HEAD.

This commit ensures the MR target is always up-to-date