provisioning: do not split --set arguments by comma
With StringSliceVarP, a comma in an argument causes it to be interpreted
as multiple arguments. This makes it difficult to include commas in the
values themselves. Use StringArrayVarP instead.

Without this change, this was necessary:
--set '"values.RepositoryPackages=exxe\,drbd-utils"'

Now, this works:
--set 'values.RepositoryPackages=exxe\,drbd-utils'

Note that the comma still needs to be escaped due to the use of
strvals.ParseInto. However, we no longer need the additional quotes to
get through StringSliceVarP.