Bind working directory in containers
Give containers read-only access to the virter working directory.

The chosen target path is '/virter/workspace', as this is unlikely
to collide with any standard image (and can be seen as a kind-of
namespacing).

ReadOnly binds were chosen since we already have a mechanism to get
files out of the container in ".copy"