Add network management commands
This commit allows virter to create new virtual networks, optionally
with DHCP and NAT. It also allows adding interfaces when creating VMs,
similar to the way additional disks can be provisioned.

The following new commands are introduced:
* network ls, showing all currently started networks
* network add, adding a new virtual network
* network rm, removing an existing network
* network list-attached, listing VMs attached to a network

The following commands are updated:
* vm run: added the `--nic` flag to add network interfaces (like `--disk`)

The main motivation for adding this was work on a multinode openstack CI
system, which work best if openstack has complete control over a single
interface.