CLI: Make advanced arg parsing reusable
The --disk flag is special and uses go tags and some reflection to
do the actual argument parsing. This is useful for other flags
that will be added in the near future.

As such the actual magic of splitting the argument and parsing it
into a tagged struct has been extracted into a new package. The
tag mini-language has been cleaned up a bit to bring it in line
with similar tags such as `json` and `toml`.

The new parser supports filling the struct directly from a string
without having to manually implement conversion from a map. It also
supports any field type implementing encoding.TextUnmarshaler.