Manage virter images like docker images
Allow pushing & pulling to docker registries. To support this, instead
of using the VM names directly as image names, images are stored in
(mostly immutable) layers that are addressed by their sha256 digest.

A VM attached volume looks like
* A working vol, named `virter:work:<vmname>`
* One or more backing volumes, named `virter:layer:sha256:<hash>`

A virter image looks like
* An (empty) tag volume, named `virter:tag:<tag>`
* One or more backing volumes, named `virter:layer:sha256:<hash>`

To create a volume from an image, just the first layer of the image
volume has to be referenced as backing store.

To create a new layer, the content of the working volume is hashed
and copied to a new volume following the above naming scheme.

In order to prevent the storage pool location to leak into the volume
data (qcow2 images store a filepath to the backing device), all backing
devices are referenced as "./$name".

Finally, to push, a manifest is created, referencing all layers needed
for the image. The layers are compressed, since docker registries expected
gzipped content.

On pull, only those layers not available locally are fetched from the registry.