fetch storage pool and network on virter init
Basically any virter command needs to look up the provision storage pool,
the provision network or both. Instead of storing only the name in the
virter struct, try to lookup both at virter initialization.

Note that virter will still initialize even if the pool or network could
not be found. In that case only a warning is printed.