Work around potential deadlock in layer clone
There seems to be a potential deadlock in the libvirt client library when
Downloading and Uploading a volume at the same time. This happens when we
convert a VM layer into a persistent image layer.

As a workaround, we first finish the volume download by writing to a temporary
file before starting the upload process.