Add option to push provisioned images
Adds three new flags to virter image build

--push automatically pushes images to the container registry indicated by
the target image name. It implements a primitive caching strategy by
comparing the "History" attribute of the remote image (if any) with
the value if the `--build-id` flag. If they match and both existing target
as well as current "image build" invocation use the same base image, actual
provisioning is skipped entirely.

--build-id is the cache key used for the primitive caching exaplained above.
It is the users responsibility to come up with a value that will change
whenever the image should be rebuild.

--no-cache disables the history check mentioned above and will always
rebuild the image before pushing.

This was mostly implemented because:
* we can
* in CI pipelines we might want to rebuild base images only when necessary.