vm exec: use independent contexts for cleanup
The overtime context allowed the cleanup steps to run until the parent
context was canceled plus a timeout. We don't really need this
complexity. The overtime timeout needs to be sufficient for the cleanup
anyway, so allowing more time in some situations doesn't bring much
benefit.

This simpler solution only allows a fixed time regardless of whether the
container terminated normally or was stopped due to the context being
canceled.

Importantly, the stop and remove cleanup steps now have independent
timeouts. Previously, if the overtime timeout triggered while stopping,
no attempt to remove the container would be made, because the context
was already canceled before "Remove" was called.