refactor: PingSSH -> WaitVmReady
We want to know when a VM can be safely accessed. This might happen
only long after sshd was configured. Since we no longer wait just
for SSH, we rename the wrapping function, so we can safely change
implementation details without confusing function consumers.

To this end, renamed:
* virter.PingSSH -> virter.WaitVmReady
* virter.SSHPingConfig -> virter.VmReadyConfig

We leave the config values related to VM readiness unchanged, since
there does not seem to be an easily implemented and backwards compatible
way of renaming those.