Add option to set pull policy for container images
Allow setting a pull policy for container images during provisioning.
The policy can be set on a per-step basis, or as a blanket override via
--container-pull-policy on all commands that can execute provision steps.

The option takes the same values as the existing --pull-policy flag, namely

* Always
* IfNotExist
* Never

They should be pretty self-describing.

To avoid confusion, this also marks the --pull-policy option deprecated.
Instead a new --vm-pull-policy option was added, which makes it clear which
option control which kinds of images.