use feature detection to determine argument order
Libvirt's network update had the argument order swapped before version 7.2.0.
We implemented a fix by checking the libvirt version (see 54a13e4fb2) and
swapping the arguments if an older libvirt version was detected.

The most recent libvirt on RHEL 8 and clones has the fix back-ported, i.e. it
silently switched the argument order for all clients.

There is actually a way to detect the presence of the fix by checking
the connection features. So instead of checking the version, we check
for the presence of that feature to determine if arguments need to be
swapped.