cloud-config: write network-config
In order to configure multiple IP addresses for different network
devices, we need to tell cloud-init to enable DHCP on for those devices.

This is easier said than done, as the different distributions we use:
* Use netplan, NetworkManager and ifcfg scripts
* Use different network device names
* Have inconsistent behaviour when handling "dhcp=false" vs the default
  (which it claims is also "false")

The only consistent behaviour between all these different distributions is:
* If no network-config is found, cloud-init will sequentially configure the
  devices  to use DHCP, meaning we configure our "access" network first
* If all attached networks run DHCP, we can tell cloud-init to configure
  all devices through dhcp.

All other cases (such as non-dhcp-enabled networks) result in some distribution
or another to time out waiting in the "network-online.target" phase of system
start up. These issues can be fixed on a per-distribution basis, but I could
not find a cloud-init solution that works for _all_ distributions at the
same time.