tests: remove broken checks
The tests used incorrect syntax to check for the presence of systemctl, which
meant all futher commands were never executed. Since we no longer test on
non-systemd systems, we can fix the check and error out if no systemctl is
found.

Also, if the system is using ifup/ifdown or similar configuration, the init
targets for networking might fail when a network-config is present for
cloud-init. As this is always the case in our e2e tests, just reset
those targets.

Apparmor service also often fails to start on certain systems, so we
just remove it for our e2e tests in the provisioning steps.

We also exclude debian-9 (Stretch) from e2e tests: cloud-init is so
old that it fails to parse the network-config v2, stopping the whole
initialization. We don't test or build against that anymore, so
it should be fine to exclude.