docker: create /etc/ssh/ssh_config.virter file
To provide remote user names for ssh and scp, a config file is mounted
into the container. This is useful, for instance, for Windows targets.

The config file is not mounted to /root/.ssh/config because then ssh
would complain about the wrong ownership of the config file when virter
is not run as root.

In order to use this config file, one needs to pass it to the ssh
commands. That means, add

        -F /etc/ssh/ssh_config.virter

to your ssh calls and do not include any explicit root@ user specification
so that the user name from the config file is used.

Alternatively, copy the config file with the appropriate ownership to
/root/.ssh/config from within the container before using ssh.