ci: run tests only for merge requests
This removes the dependency of the sync-images job on a successful build
and test run.

Without this change, the sync-images may not run because the e2e-test
job fails. It may be that the e2e-test job is failing because the images
are not synced. Break this circular dependency.

Also allow the pipeline to run when branches other than master are
scheduled.