enable ipv6 networks
Configure VMs to expect a DHCPv6 server to be running if the
virtual network supports it. Additionally extend the network
commands to also provisioning networks with DHCPv6 enabled.

For all virter operations we still require IPv4, as a fixed
assignement from VM to IPv6 is hard to impossible because MAC
addresses can't be used in DHCPv6.