add version field to provisioning file
Before committing to a stable interface, we want to make sure that the
provisioning file format can be safely evolved. To allow this, add a
version field and reject any file with a version != 0.

Since the default value of an unset toml integer field is 0, this is a
no-op until we increase the CurrentProvisionFileVersion.