virter/image: pass storage pool arguments
Refactoring in preparation for the next commit. Change the image and
layer logic so that an arbitrary storage pool could be used, not just
the provisionStoragePool as is currently the case.

Note that this specifically only adds support for "non-default" storage
pools for "dynamic" layers, i.e. not the boot image or the cloud-init
data image. These still always have to be in the default storage pool.