virter: pass-through exit code from provisioning
When a provisioning step fails, it may be helpful to the caller of
virter to know the exit code of the script.

So, if an SSH or container provisioning step fails with an exit code,
pass it through to the CLI, exiting from the virter process with the
same exit code.