network: fix calculation of DHCP range for /16 networks
When generating DHCP ranges for /16 networks, we are right at the limits
of what libvirt supports. We try to limit the number of hosts to at most
2^16-1 in case of large subnets.

However, the check was incorrect, as for a /16 network we would define the end
range as being .255.255, i.e. the broadcast address. So when we have a /16
network, we are still in the default case of having to decrement the "last" IP
by one.