cmd: add wait-ready command to check VM SSH accessibility
This works very similarly to the --wait-ssh flag on the "vm run"
command (and uses the same logic under the hood).
The difference is that this can be used independently of the VM
creation process, for example when rebooting an already running VM and
waiting for it to come back online.